pyproject.toml
[project]
name = "fashion-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djoser>=2.3.3",
    "python-dotenv>=1.1.1",
    "mysqlclient>=2.2.0",
    "uvicorn[standard]>=0.30.0",
    "django-cors-headers>=4.4.0",
    "requests>=2.32.0",
]
